refactor: split order component
